Move applyLifecycleHooks logic into helper function
